Divided for geo_info and run files
